requirements.txt
alembic==1.13.1
apache-airflow==2.9.1
apache-airflow-providers-cncf-kubernetes==8.2.0
docker==7.0.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.22.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.16.0
json-merge-patch==0.2
jsonschema==4.22.0
pandas==2.0.3
pendulum==3.0.0
python-dateutil==2.9.0.post0
requests==2.31.0
SQLAlchemy==1.4.52
security==1.3.1
